requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.9.0
pytesseract==0.3.10
opencv-python-headless==4.8.0.76
Pillow==10.1.0
pdfplumber==0.10.3
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1
numpy==1.24.0
